+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/tokenaccess/TokenWriteAccess.java
@@ -1,7 +1,7 @@
-package de.aaaaaaah.velcom.backend.access;
+package de.aaaaaaah.velcom.backend.newaccess.tokenaccess;
 
-import de.aaaaaaah.velcom.backend.access.entities.AuthToken;
 import de.aaaaaaah.velcom.backend.newaccess.repoaccess.entities.RepoId;
+import de.aaaaaaah.velcom.backend.newaccess.tokenaccess.entities.AuthToken;
 import de.aaaaaaah.velcom.backend.storage.db.DatabaseStorage;
 import java.util.Objects;
 import javax.annotation.Nullable;
