+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/tokenaccess/entities/AuthToken.java
@@ -1,4 +1,4 @@
-package de.aaaaaaah.velcom.backend.access.entities;
+package de.aaaaaaah.velcom.backend.newaccess.tokenaccess.entities;
 
 import java.util.Objects;
