+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/tokenaccess/hashalgorithm/HashAlgorithm.java
@@ -1,7 +1,7 @@
-package de.aaaaaaah.velcom.backend.access.hashalgorithm;
+package de.aaaaaaah.velcom.backend.newaccess.tokenaccess.hashalgorithm;
 
 
-import de.aaaaaaah.velcom.backend.access.entities.AuthToken;
+import de.aaaaaaah.velcom.backend.newaccess.tokenaccess.entities.AuthToken;
 
 /**
  * Represents a certain hashing algorithm used for storing tokens.
