+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/tokenaccess/hashalgorithm/V1Argon2Algorithm.java
@@ -1,6 +1,6 @@
-package de.aaaaaaah.velcom.backend.access.hashalgorithm;
+package de.aaaaaaah.velcom.backend.newaccess.tokenaccess.hashalgorithm;
 
-import de.aaaaaaah.velcom.backend.access.entities.AuthToken;
+import de.aaaaaaah.velcom.backend.newaccess.tokenaccess.entities.AuthToken;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import de.mkammerer.argon2.Argon2Factory.Argon2Types;
