+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/restapi/authentication/RepoAuthenticator.java
@@ -1,8 +1,8 @@
 package de.aaaaaaah.velcom.backend.restapi.authentication;
 
-import de.aaaaaaah.velcom.backend.access.TokenReadAccess;
-import de.aaaaaaah.velcom.backend.access.entities.AuthToken;
 import de.aaaaaaah.velcom.backend.newaccess.repoaccess.entities.RepoId;
+import de.aaaaaaah.velcom.backend.newaccess.tokenaccess.TokenReadAccess;
+import de.aaaaaaah.velcom.backend.newaccess.tokenaccess.entities.AuthToken;
 import io.dropwizard.auth.Authenticator;
 import io.dropwizard.auth.basic.BasicCredentials;
 import java.util.Optional;
